refactor(caa dims): clean up index.ts
Final touches of refactoring.
